Update README.md
This program takes a string as user input and reverse that string. This program also allows the user to define a set of characters that are restricted, and therefore, cannot be swapped. In a case of this, the pointers will skip over the restricted character, moving onto the next one. Both the pointers will always check before swapping characters, if either of them are restricted, if so, the pointer that is currently on a restricted character, moves to the next character and re-checks. The swap is only performed, once both characters are not restricted.